Added implementation of insertion sort algorithm
Insertion sorts works in a similar way to bubble sort, the rigorous way is to have an output list that you fill with the values being sorted along the process, but I made it so that the output array is the same as the input one. 
The algorithm itself is straightforward ; pick the first value of the array and consider it sorted. Then, for every value of the unsorted array, traverse it in reverse and swap the values if the nth value is lower than the nth-1 value.